# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_dataset_v1_datasets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_v1_datasets_post",
        "type": "object"
      },
      "DatasetRecord": {
        "description": "A stored dataset with its report attached -- including a dataset that\nfailed validation, whose report is the reason it was kept.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "created_at",
          "status"
        ],
        "title": "DatasetRecord",
        "type": "object"
      },
      "DatasetReport": {
        "description": "What validation says about a dataset. The same dict\n`temper_core.validation.Report.to_dict()` produces, published typed.",
        "properties": {
          "enable_thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Thinking"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Errors",
            "type": "array"
          },
          "preview": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Preview",
            "type": "array"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "schema_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Type"
          },
          "usable_rows": {
            "title": "Usable Rows",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "valid",
          "row_count",
          "usable_rows",
          "errors",
          "warnings",
          "preview"
        ],
        "title": "DatasetReport",
        "type": "object"
      },
      "DatasetUploaded": {
        "description": "The response to a successful upload: the report plus its identity.",
        "properties": {
          "enable_thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Thinking"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Errors",
            "type": "array"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preview": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Preview",
            "type": "array"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "schema_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Type"
          },
          "usable_rows": {
            "title": "Usable Rows",
            "type": "integer"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "valid",
          "row_count",
          "usable_rows",
          "errors",
          "warnings",
          "preview",
          "id",
          "filename"
        ],
        "title": "DatasetUploaded",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobRequest": {
        "properties": {
          "base_model": {
            "default": "qwen3-4b",
            "title": "Base Model",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "hyperparameters": {
            "additionalProperties": true,
            "title": "Hyperparameters",
            "type": "object"
          }
        },
        "required": [
          "dataset_id"
        ],
        "title": "JobRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "description": "One problem, named where it is: line number when one applies.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Fine-tuning platform. Dataset in, adapter out.",
    "title": "Temper",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/datasets": {
      "get": {
        "operationId": "list_datasets_v1_datasets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "datasets"
        ]
      },
      "post": {
        "description": "Upload and validate a JSONL dataset.\n\n**Deliberately a sync handler.** It was `async def`, which ran the\nCPU-bound validation on the event loop and froze every other request for\nas long as it ran -- measured at roughly 80 ms per megabyte, so a 200 MB\nupload blocked the whole server for ~16 seconds. Declared `def` instead,\nFastAPI dispatches it to its worker pool and a long validation blocks only\nits own request. Regression-tested by\n`test_upload_limits.test_large_upload_does_not_block_concurrent_requests`.\n\nStorage and validation live in `datasets.store_and_validate`, shared with\nthe browser's upload form, so the two surfaces cannot drift apart.\n\nDatasets over `config.MAX_DATASET_BYTES` are refused before validation --\nan unbounded upload fails as an out-of-memory crash rather than a typed\nerror, which is worse for the user and for the process.",
        "operationId": "upload_dataset_v1_datasets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_v1_datasets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetUploaded"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/v1/datasets/{dataset_id}": {
      "get": {
        "operationId": "get_dataset_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/v1/jobs": {
      "get": {
        "operationId": "list_jobs_v1_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Launch a fine-tuning job.\n\nThe whole creation path lives in `api.jobs.create`, shared with the\nbrowser's launch form so the two surfaces cannot drift apart. The\nhyperparameters are frozen into the job row there: a run's spec is\nimmutable once launched, so a later change to a default cannot\nretroactively alter what a finished run claims.",
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/adapter": {
      "get": {
        "operationId": "download_adapter_v1_jobs__job_id__adapter_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Adapter",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Ask a running job to stop. Destructive, and says so.\n\nThe request records the user's decision; the thread running the job sees it\nwithin a poll interval, destroys the machine and ends the job `cancelled`.\nSo this returns the state the job is in *now*, which is usually still a\nworking one \u2014 reporting `cancelled` here would claim a teardown that has\nnot happened yet, and teardown is the whole point of cancelling.\n\nRepeating it succeeds quietly: a double-clicked button is not an error.\nCancelling a job that has already ended is refused with a stable code,\nbecause nothing was undone and the user should not think otherwise.",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Durable event log. `after` is the last event id the client holds.\n\nPolling against a monotonic id rather than streaming: a reconnecting client\ncatches up from the database instead of losing whatever happened while it\nwas away.",
        "operationId": "get_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/models": {
      "get": {
        "description": "The curated base-model catalog.\n\nAn allow-list, not a limitation: detection of an arbitrary architecture is\neasy, but *support* means testing its chat template, tokenizer quirks and\npacking compatibility. This list is a promise about what has been tested.",
        "operationId": "list_models_v1_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Models",
        "tags": [
          "catalog"
        ]
      }
    }
  }
}